+++ packages/xrpl/snippets/src/paths.ts
@@ -5,7 +5,9 @@ const client = new Client('wss://s.altnet.rippletest.net:51233')
 async function createTxWithPaths(): Promise<void> {
   await client.connect()
 
-  const { wallet } = await client.fundWallet()
+  const { wallet } = await client.fundWallet(null, {
+    usageContext: 'code snippets',
+  })
   const destination_account = 'rKT4JX4cCof6LcDYRz8o3rGRu7qxzZ2Zwj'
   const destination_amount = {
     value: '0.001',