+++ packages/xrpl/snippets/src/setRegularKey.ts
@@ -8,9 +8,15 @@ const client = new Client('wss://s.altnet.rippletest.net:51233')
  */
 async function setRegularKey(): Promise<void> {
   await client.connect()
-  const { wallet: wallet1 } = await client.fundWallet()
-  const { wallet: wallet2 } = await client.fundWallet()
-  const { wallet: regularKeyWallet } = await client.fundWallet()
+  const { wallet: wallet1 } = await client.fundWallet(null, {
+    usageContext: 'code snippets',
+  })
+  const { wallet: wallet2 } = await client.fundWallet(null, {
+    usageContext: 'code snippets',
+  })
+  const { wallet: regularKeyWallet } = await client.fundWallet(null, {
+    usageContext: 'code snippets',
+  })
 
   console.log('Balances before payment')
   console.log(await client.getXrpBalance(wallet1.classicAddress))