+++ packages/xrpl/src/index.ts
@@ -1,5 +1,5 @@
 // Broadcast client is experimental
-export { default as BroadcastClient } from './client/BroadcastClient'
+export { BroadcastClient } from './client/BroadcastClient'
 
 export { Client, ClientOptions } from './client'
 
@@ -7,9 +7,11 @@ export * from './models'
 
 export * from './utils'
 
+export { default as ECDSA } from './ECDSA'
+
 export * from './errors'
 
-export { default as Wallet } from './Wallet'
+export { Wallet } from './Wallet'
 
 export { keyToRFC1751Mnemonic, rfc1751MnemonicToKey } from './Wallet/rfc1751'