+++ packages/xrpl/src/models/methods/index.ts
@@ -1,24 +1,59 @@
+/* eslint-disable max-lines -- There is a lot to export */
 import {
   AccountChannelsRequest,
   AccountChannelsResponse,
+  Channel,
 } from './accountChannels'
 import {
   AccountCurrenciesRequest,
   AccountCurrenciesResponse,
 } from './accountCurrencies'
-import { AccountInfoRequest, AccountInfoResponse } from './accountInfo'
-import { AccountLinesRequest, AccountLinesResponse } from './accountLines'
-import { AccountNFTsRequest, AccountNFTsResponse } from './accountNFTs'
-import { AccountObjectsRequest, AccountObjectsResponse } from './accountObjects'
+import {
+  AccountInfoAccountFlags,
+  AccountInfoRequest,
+  AccountInfoResponse,
+  AccountQueueData,
+  AccountQueueTransaction,
+} from './accountInfo'
+import {
+  AccountLinesRequest,
+  AccountLinesResponse,
+  AccountLinesTrustline,
+} from './accountLines'
+import {
+  AccountNFToken,
+  AccountNFTsRequest,
+  AccountNFTsResponse,
+} from './accountNFTs'
+import {
+  AccountObject,
+  AccountObjectsRequest,
+  AccountObjectsResponse,
+  AccountObjectType,
+} from './accountObjects'
 import {
   AccountOffer,
   AccountOffersRequest,
   AccountOffersResponse,
 } from './accountOffers'
-import { AccountTxRequest, AccountTxResponse } from './accountTx'
+import {
+  AccountTxRequest,
+  AccountTxResponse,
+  AccountTxTransaction,
+} from './accountTx'
 import { AMMInfoRequest, AMMInfoResponse } from './ammInfo'
-import { ErrorResponse } from './baseMethod'
-import { BookOffersRequest, BookOffer, BookOffersResponse } from './bookOffers'
+import {
+  BaseRequest,
+  BaseResponse,
+  ErrorResponse,
+  ResponseWarning,
+} from './baseMethod'
+import {
+  BookOffersRequest,
+  BookOffer,
+  BookOffersResponse,
+  BookOfferCurrency,
+} from './bookOffers'
 import { ChannelVerifyRequest, ChannelVerifyResponse } from './channelVerify'
 import {
   DepositAuthorizedRequest,
@@ -26,17 +61,34 @@ import {
 } from './depositAuthorized'
 import { FeeRequest, FeeResponse } from './fee'
 import {
+  GatewayBalance,
   GatewayBalancesRequest,
   GatewayBalancesResponse,
 } from './gatewayBalances'
-import { LedgerRequest, LedgerResponse } from './ledger'
+import {
+  LedgerBinary,
+  LedgerModifiedOfferCreateTransaction,
+  LedgerQueueData,
+  LedgerRequest,
+  LedgerResponse,
+} from './ledger'
 import { LedgerClosedRequest, LedgerClosedResponse } from './ledgerClosed'
 import { LedgerCurrentRequest, LedgerCurrentResponse } from './ledgerCurrent'
-import { LedgerDataRequest, LedgerDataResponse } from './ledgerData'
+import {
+  LedgerDataBinaryLedgerEntry,
+  LedgerDataLabeledLedgerEntry,
+  LedgerDataLedgerState,
+  LedgerDataRequest,
+  LedgerDataResponse,
+} from './ledgerData'
 import { LedgerEntryRequest, LedgerEntryResponse } from './ledgerEntry'
 import { ManifestRequest, ManifestResponse } from './manifest'
 import { NFTBuyOffersRequest, NFTBuyOffersResponse } from './nftBuyOffers'
-import { NFTHistoryRequest, NFTHistoryResponse } from './nftHistory'
+import {
+  NFTHistoryRequest,
+  NFTHistoryResponse,
+  NFTHistoryTransaction,
+} from './nftHistory'
 import { NFTInfoRequest, NFTInfoResponse } from './nftInfo'
 import { NFTSellOffersRequest, NFTSellOffersResponse } from './nftSellOffers'
 import { NoRippleCheckRequest, NoRippleCheckResponse } from './norippleCheck'
@@ -46,11 +98,24 @@ import {
   PathFindCreateRequest,
   PathFindStatusRequest,
   PathFindResponse,
+  PathFindPathOption,
 } from './pathFind'
 import { PingRequest, PingResponse } from './ping'
 import { RandomRequest, RandomResponse } from './random'
-import { RipplePathFindRequest, RipplePathFindResponse } from './ripplePathFind'
-import { ServerInfoRequest, ServerInfoResponse } from './serverInfo'
+import {
+  RipplePathFindPathOption,
+  RipplePathFindRequest,
+  RipplePathFindResponse,
+  SourceCurrencyAmount,
+} from './ripplePathFind'
+import {
+  JobType,
+  ServerInfoRequest,
+  ServerInfoResponse,
+  ServerState,
+  StateAccounting,
+  StateAccountingFinal,
+} from './serverInfo'
 import { ServerStateRequest, ServerStateResponse } from './serverState'
 import { SubmitRequest, SubmitResponse } from './submit'
 import {
@@ -58,12 +123,15 @@ import {
   SubmitMultisignedResponse,
 } from './submitMultisigned'
 import {
+  BooksSnapshot,
   ConsensusStream,
   LedgerStream,
+  LedgerStreamResponse,
   OrderBookStream,
   PathFindStream,
   PeerStatusStream,
   Stream,
+  SubscribeBook,
   SubscribeRequest,
   SubscribeResponse,
   TransactionStream,
@@ -74,7 +142,11 @@ import {
   TransactionEntryResponse,
 } from './transactionEntry'
 import { TxRequest, TxResponse } from './tx'
-import { UnsubscribeRequest, UnsubscribeResponse } from './unsubscribe'
+import {
+  UnsubscribeBook,
+  UnsubscribeRequest,
+  UnsubscribeResponse,
+} from './unsubscribe'
 /**
  * @category Requests
  */
@@ -180,19 +252,31 @@ type Response =
   | NFTHistoryResponse
 
 export {
+  // Allow users to define their own requests and responses.  This is useful for releasing experimental versions
+  BaseRequest,
+  BaseResponse,
   Request,
   Response,
-  // account methods
+  ResponseWarning,
+  // account methods with types
+  Channel,
   AccountChannelsRequest,
   AccountChannelsResponse,
   AccountCurrenciesRequest,
   AccountCurrenciesResponse,
+  AccountInfoAccountFlags,
   AccountInfoRequest,
   AccountInfoResponse,
+  AccountQueueData,
+  AccountQueueTransaction,
   AccountLinesRequest,
   AccountLinesResponse,
+  AccountLinesTrustline,
+  AccountNFToken,
   AccountNFTsRequest,
   AccountNFTsResponse,
+  AccountObject,
+  AccountObjectType,
   AccountObjectsRequest,
   AccountObjectsResponse,
   AccountOffer,
@@ -200,8 +284,10 @@ export {
   AccountOffersResponse,
   AccountTxRequest,
   AccountTxResponse,
+  AccountTxTransaction,
   AMMInfoRequest,
   AMMInfoResponse,
+  GatewayBalance,
   GatewayBalancesRequest,
   GatewayBalancesResponse,
   NoRippleCheckRequest,
@@ -209,15 +295,21 @@ export {
   // ledger methods
   LedgerRequest,
   LedgerResponse,
+  LedgerQueueData,
+  LedgerBinary,
+  LedgerModifiedOfferCreateTransaction,
   LedgerClosedRequest,
   LedgerClosedResponse,
   LedgerCurrentRequest,
   LedgerCurrentResponse,
   LedgerDataRequest,
+  LedgerDataLabeledLedgerEntry,
+  LedgerDataBinaryLedgerEntry,
   LedgerDataResponse,
+  LedgerDataLedgerState,
   LedgerEntryRequest,
   LedgerEntryResponse,
-  // transaction methods
+  // transaction methods with types
   SubmitRequest,
   SubmitResponse,
   SubmitMultisignedRequest,
@@ -226,27 +318,34 @@ export {
   TransactionEntryResponse,
   TxRequest,
   TxResponse,
-  // path and order book methods
+  // path and order book methods with types
   BookOffersRequest,
   BookOffer,
+  BookOfferCurrency,
   BookOffersResponse,
   DepositAuthorizedRequest,
   DepositAuthorizedResponse,
   PathFindRequest,
   PathFindCreateRequest,
   PathFindCloseRequest,
+  PathFindPathOption,
   PathFindStatusRequest,
   PathFindResponse,
+  RipplePathFindPathOption,
   RipplePathFindRequest,
   RipplePathFindResponse,
+  SourceCurrencyAmount,
   // payment channel methods
   ChannelVerifyRequest,
   ChannelVerifyResponse,
-  // Subscribe methods/streams
+  // Subscribe methods/streams with types
   SubscribeRequest,
   SubscribeResponse,
+  SubscribeBook,
   Stream,
+  BooksSnapshot,
   LedgerStream,
+  LedgerStreamResponse,
   ValidationStream,
   TransactionStream,
   PathFindStream,
@@ -255,7 +354,8 @@ export {
   ConsensusStream,
   UnsubscribeRequest,
   UnsubscribeResponse,
-  // server info methods
+  UnsubscribeBook,
+  // server info methods with types
   FeeRequest,
   FeeResponse,
   ManifestRequest,
@@ -264,6 +364,10 @@ export {
   ServerInfoResponse,
   ServerStateRequest,
   ServerStateResponse,
+  JobType,
+  ServerState,
+  StateAccountingFinal,
+  StateAccounting,
   // utility methods
   PingRequest,
   PingResponse,
@@ -280,4 +384,5 @@ export {
   NFTInfoResponse,
   NFTHistoryRequest,
   NFTHistoryResponse,
+  NFTHistoryTransaction,
 }